Memory Management Changes
Redid a ton of code. Made it where the child process that spawns is on a
global interval instead of depending on the route. Got rid of route
caching entirely, as that could have been causing some issues with GC
not working properly. Added in a /debug route to manually refresh the
cache if need be, mostly used for science.

Changed the front page to display more "featured" streams. Temporarily
removed the feature to filter and sort through online streams. Going to
build a completely client side solution, I believe having server side
handle filtering was causing some memory issues and creating completely
un-needed objects.